Matrix Diagonal Addi-tor
I made this code in which it adds the diagonals of the matrix. It first asks the user the number of rows and columns. Then asks for the elements in the matrix. After that, it prints out the matrix and then calculates and print the sum of the primary and secondary diagonal.